+++ CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v1.0.3] PENDING
-### Fix
+## [v1.0.3] 2025-03-06
+### Fixed
 * Rename minLen to minLength and maxLen to maxLength
 * Skip validation for null values for min and max values if the column is not required.
 * Refactor parameter validations in the entity schema definition
+* Add PK validation: Must be only one PK.
+* Add unique validation
+### Added
+* Export MysqlCoreError, MysqlEntityDefinitionError, MysqlEntityValidationError and MysqlEntityNotFound
 
 ## [v1.0.2] 2025-01-28
 ### Added